+++ uwc-backend/Models/Mcp.cs
@@ -2,8 +2,8 @@ namespace Models;
 
 public class Mcp : IndexedEntity
 {
-    public double Capacity { get; set; }
-    public double CurrentLoad { get; set; }
+    public float Capacity { get; set; }
+    public float CurrentLoad { get; set; }
     public double Latitude { get; set; }
     public double Longitude { get; set; }
 }
